email-extractor/requirements.txt
groq>=0.9.0
pydantic==2.6.4
python-dotenv==1.0.1
tqdm==4.66.2
pytest==8.1.1
